chore(mobile): remove dead Vapi/Daily integration (realtime-only now)
The Vapi/custom-pipeline voice modes were already unreachable from the UI;
the only voice path is the realtime WebSocket → relay. The vendored
Daily.xcframework just bloated the IPA and produced "Upload Symbols
Failed" dSYM warnings on TestFlight.

Deleted:
- mobile/modules/expo-vapi/ (Vapi Swift bridge + vendored Daily SDK)
- mobile/scripts/download-daily-sdk.sh
- "Download Daily.co iOS SDK" step from scripts/build-ios.sh
- Daily.xcframework .gitignore entry
- vapi case + validateVapi() from lib/validate-api-key.ts
- All ExpoVapiModule listeners, latency tracking refs, and
  custom-pipeline call sites in app/(tabs)/index.tsx
- Vapi key / assistant id / default model UI + state in
  app/(tabs)/settings.tsx
- expo-vapi import in lib/sounds.ts

Added:
- mobile/modules/expo-call-sounds/ — minimal Expo module that just plays
  the bundled call-join / call-end / thinking wav files via
  AVAudioPlayer (no Vapi, no Daily, no networking). lib/sounds.ts now
  uses this. Preserves call-sound UX without dragging in the Daily SDK.

Preserved:
- voice_mode 'vapi'/'custom' → 'realtime' migration in settings.tsx so
  existing installs land on the only supported mode.
- Custom-pipeline native module (still used by Kokoro download UI).
- Realtime path, useCallSounds API, all other features.

tsc --noEmit passes (1242 files, exit 0). Grep for expo-vapi / ExpoVapi
/ Daily / @daily across mobile/{app,lib,scripts,app.config.ts,package.json}
returns nothing.

Next build needs `expo prebuild --clean` so the regenerated ios/
project drops Daily.xcframework and picks up the new expo-call-sounds
podspec.
