+++ mobile/lib/sounds.ts
@@ -1,27 +1,27 @@
-import ExpoVapiModule from '@/modules/expo-vapi'
+import ExpoCallSoundsModule from '@/modules/expo-call-sounds'
 import { useCallback, useRef } from 'react'
 
 export function useCallSounds() {
   const thinkingActive = useRef(false)
 
   const playJoin = useCallback(() => {
-    ExpoVapiModule.playSound('call-join', 0.4)
+    ExpoCallSoundsModule.playSound('call-join', 0.4)
   }, [])
 
   const playEnd = useCallback(() => {
-    ExpoVapiModule.playSound('call-end', 0.3)
+    ExpoCallSoundsModule.playSound('call-end', 0.3)
   }, [])
 
   const startThinking = useCallback(() => {
     if (thinkingActive.current) return
     thinkingActive.current = true
-    ExpoVapiModule.playSound('thinking', 0.15, true)
+    ExpoCallSoundsModule.playSound('thinking', 0.15, true)
   }, [])
 
   const stopThinking = useCallback(() => {
     if (!thinkingActive.current) return
     thinkingActive.current = false
-    ExpoVapiModule.stopSound()
+    ExpoCallSoundsModule.stopSound()
   }, [])
 
   return { playJoin, playEnd, startThinking, stopThinking }
