+++ mobile/lib/validate-api-key.ts
@@ -5,7 +5,7 @@ export type ValidationResult = {
   error?: string
 }
 
-export type Provider = 'brain' | 'elevenlabs' | 'deepgram' | 'openai_tts' | 'vapi'
+export type Provider = 'brain' | 'elevenlabs' | 'deepgram' | 'openai_tts'
 
 export async function validateApiKey(
   provider: Provider,
@@ -25,8 +25,6 @@ export async function validateApiKey(
       return validateDeepgram(apiKey)
     case 'openai_tts':
       return validateOpenAI(apiKey)
-    case 'vapi':
-      return validateVapi(apiKey)
     default:
       return { status: 'invalid', error: 'Unknown provider' }
   }
@@ -123,21 +121,3 @@ async function validateOpenAI(apiKey: string): Promise<ValidationResult> {
   }
 }
 
-async function validateVapi(apiKey: string): Promise<ValidationResult> {
-  try {
-    const response = await fetch('https://api.vapi.ai/assistant', {
-      method: 'GET',
-      headers: { Authorization: `Bearer ${apiKey}` },
-    })
-
-    if (response.status === 401 || response.status === 403) {
-      return { status: 'invalid', error: 'Invalid API key' }
-    }
-    if (response.ok) {
-      return { status: 'valid' }
-    }
-    return { status: 'invalid', error: `HTTP ${response.status}` }
-  } catch (err: any) {
-    return { status: 'invalid', error: err.message || 'Network error' }
-  }
-}
