+++ mobile/modules/expo-call-sounds/expo-module.config.json
@@ -1,6 +1,6 @@
 {
   "platforms": ["apple"],
   "apple": {
-    "modules": ["ExpoVapiModule"]
+    "modules": ["ExpoCallSoundsModule"]
   }
 }
