+++ mobile/modules/expo-call-sounds/src/ExpoCallSoundsModule.ts
@@ -1,0 +1,8 @@
+import { NativeModule, requireNativeModule } from 'expo'
+
+declare class ExpoCallSoundsModule extends NativeModule {
+  playSound(soundName: string, volume: number, loop?: boolean): void
+  stopSound(): void
+}
+
+export default requireNativeModule<ExpoCallSoundsModule>('ExpoCallSounds')
