--- mobile/modules/expo-vapi/index.ts
@@ -1,3 +1,0 @@
-export { default } from './src/ExpoVapiModule'
-export * from './src/ExpoVapi.types'
-export { useVapi } from './src/useVapi'
