--- mobile/modules/expo-vapi/ios/Vapi/Constants/Constants.swift
@@ -1,13 +1,0 @@
-//
-//  Constants.swift
-//
-//
-//  Created by CITMark on 03/09/2024.
-//
-
-import Foundation
-
-extension String {
-    /// This value represents the `username` of the remote `Daily.Participant` for the `Vapi Speaker`
-    static let remoteParticipantVapiSpeaker = "Vapi Speaker"
-}
