--- mobile/modules/expo-vapi/ios/Vapi/Models/AssistantStarted.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct AssistantStarted: Codable {
-    public let newAssistant: JSONValue
-}
