--- mobile/modules/expo-vapi/ios/Vapi/Models/CallDeleteFailed.swift
@@ -1,3 +1,0 @@
-import Foundation
-
-public struct CallDeleteFailed: Codable {}
