--- mobile/modules/expo-vapi/ios/Vapi/Models/CallDeleted.swift
@@ -1,3 +1,0 @@
-import Foundation
-
-public struct CallDeleted: Codable {}
