--- mobile/modules/expo-vapi/ios/Vapi/Models/ChatCreated.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct ChatCreated: Codable {
-    public let chat: JSONValue
-}
