--- mobile/modules/expo-vapi/ios/Vapi/Models/ChatDeleted.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct ChatDeleted: Codable {
-    public let chat: JSONValue
-}
