--- mobile/modules/expo-vapi/ios/Vapi/Models/LanguageChangeDetected.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct LanguageChangeDetected: Codable {
-    public let language: String
-}
