--- mobile/modules/expo-vapi/ios/Vapi/Models/Metadata.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct Metadata: Codable {
-    public let metadata: String
-}
