--- mobile/modules/expo-vapi/ios/Vapi/Models/ModelOutput.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct ModelOutput: Codable {
-    public let output: String
-}
