--- mobile/modules/expo-vapi/ios/Vapi/Models/SessionCreated.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct SessionCreated: Codable {
-    public let session: JSONValue
-}
