--- mobile/modules/expo-vapi/ios/Vapi/Models/SessionDeleted.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct SessionDeleted: Codable {
-    public let session: JSONValue
-}
