--- mobile/modules/expo-vapi/ios/Vapi/Models/SessionUpdated.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct SessionUpdated: Codable {
-    public let session: JSONValue
-}
