--- mobile/modules/expo-vapi/ios/Vapi/Models/StatusUpdate.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct StatusUpdate: Codable {
-    public let status: String
-}
