--- mobile/modules/expo-vapi/ios/Vapi/Models/ToolCalls.swift
@@ -1,6 +1,0 @@
-import Foundation
-
-public struct ToolCalls: Codable {
-    public let toolWithToolCallList: [JSONValue]
-    public let toolCallList: [JSONValue]
-}
