--- mobile/modules/expo-vapi/ios/Vapi/Models/ToolCallsResult.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct ToolCallsResult: Codable {
-    public let toolCallResult: JSONValue
-}
