--- mobile/modules/expo-vapi/ios/Vapi/Models/UserInterrupted.swift
@@ -1,10 +1,0 @@
-//
-//  UserInterrupted.swift
-//  Vapi
-//
-//  Created by Abizar Bagasrawala on 11/1/24.
-//
-
-import Foundation
-
-public struct UserInterrupted: Codable {}
