--- mobile/modules/expo-vapi/ios/Vapi/Models/VoiceInput.swift
@@ -1,12 +1,0 @@
-//
-//  VoiceInput.swift
-//  Vapi
-//
-//  Created by Abizar Bagasrawala on 11/1/24.
-//
-
-import Foundation
-
-public struct VoiceInput: Codable {
-    public let input: String
-}
