--- mobile/modules/expo-vapi/ios/Vapi/Models/WorkflowNodeStarted.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-public struct WorkflowNodeStarted: Codable {
-    public let node: JSONValue
-}
